Cast API timeSpentBackend to an int
Fixes {"timeSpentBackend":"Expected integer, but recieved double"} Avro
encoding error.

Bug: T108618

+++ includes/api/ApiMain.php
@@ -1367,7 +1367,7 @@ class ApiMain extends ApiBase {
 			'ip' => $request->getIP(),
 			'userAgent' => $this->getUserAgent(),
 			'wiki' => wfWikiID(),
-			'timeSpentBackend' => round( $time * 1000 ),
+			'timeSpentBackend' => (int) round( $time * 1000 ),
 			'hadError' => $e !== null,
 			'errorCodes' => [],
 			'params' => [],
